feat: Request App Tracking Transparency authorization on launch (#6)
* feat: Request App Tracking Transparency authorization on launch

Implements the App Tracking Transparency (ATT) framework to request user authorization for accessing the device's advertising identifier (IDFA) on iOS.

This change introduces the app_tracking_transparency package and converts the SoloScriptedApp widget into a StatefulWidget. This allows for the authorization request to be triggered within the initState method, ensuring that the permission dialog is presented to the user when the application first initializes.

The request is only made if the authorization status is notDetermined, preventing repeated prompts to the user. This is a necessary step for compliance with Apple's privacy policies on iOS 14.5 and later.

Signed-off-by: Mustafa Kara <mkara@soloscripted.com>

* fix:dependencies

Signed-off-by: Mustafa Kara <mkara@soloscripted.com>

* fix:formatting

Signed-off-by: Mustafa Kara <mkara@soloscripted.com>

* fix:formatting

Signed-off-by: Mustafa Kara <mkara@soloscripted.com>

---------

Signed-off-by: Mustafa Kara <mkara@soloscripted.com>